+++ R/identify_rows_dataframe.R
@@ -87,11 +87,6 @@ identify_rows.data.frame <- function(data, vars, match = c("in", "start", "regex
   if (verbose) {
     result_vals <- dt[, unlist(.SD), .SDcols = vars] %>% unique()
 
-    # cat(
-    #   "\nIdentify records with condition(s):",
-    #   "\n - where", ifelse(if_all & length(vars) > 1, "all of the", ifelse(length(vars) > 1, "at least one of the", "the")), paste0(vars, collapse = ", "), "column(s) in each record",
-    #   "\n   - contains a value", match_msg, match_str, "\n"
-    # )
     rlang::inform(c(
       "i" = "Identify records with condition(s):",
       "*" = glue::glue('where {ifelse(if_all & length(vars) > 1, "all of the", ifelse(length(vars) > 1, "at least one of the", "the"))} {paste0(vars, collapse = ", ")} column(s) in each record'),